+++ src/apprun-html.ts
@@ -1,8 +1,8 @@
-import app, { Component } from './apprun'
+import app, { Component, View, Action, Update, on, update } from './apprun'
 import { createElement, render } from './vdom-html';
 
 app.createElement = createElement;
 app.render = render;
 
 export default app;
-export { Component };
+export { Component, View, Action, Update, on, update };